requirements.txt
flask
pytest
openai
python-dotenv
